requirements.txt
ldap3
PyYAML
colorama
impacket==0.9.23
pyOpenSSL
